LayoutTests/http/tests/navigation/history-back-across-form-submission-to-fragment.html

<body>
<script>
if (window.testRunner) {
  testRunner.dumpAsText();
  testRunner.waitUntilDone();
}
// This test confirms that when we submit a form to a reference fragment, that
// navigating back, causes the page to load again.  We want to ensure that the
// form submission created a new document and that when navigating back, we
// don't confuse this as a same-document navigation.
onload = function() {
  sessionStorage.stage = 1;
  location = "resources/submit-to-fragment.pl";
}
</script>
</body>
