LayoutTests/http/tests/navigation/history-back-across-form-submission-to-fragment-expected.txt

main frame - has 1 onunload handler(s)
ALERT: stage: 1
main frame - has 1 onunload handler(s)
ALERT: stage: 2
main frame - has 1 onunload handler(s)
ALERT: stage: 3

PASS
